+++ locallib.php
@@ -85,6 +85,10 @@ function local_boostnavigation_build_custom_nodes($customnodes, navigation_node 
     $lastparentnode = null;
     $lastparentnodevisible = false;
 
+    // Initialize variables for marking the first custom bottom parent nodes.
+    $firstcustombuttomusersparentnodeseen = false;
+    $firstcustombuttomadminsparentnodeseen = false;
+
     // Initialize variables for remembering the node keys for collapsing.
     $collapsenodesforjs = array();
     $collapselastparentprepared = false;
@@ -318,13 +322,20 @@ function local_boostnavigation_build_custom_nodes($customnodes, navigation_node 
                 if ($collapse) {
                     // Remember that we haven't prepared collapsing yet for this parent node.
                     $collapselastparentprepared = false;
+                }
 
-                    // If the node shouldn't be collapsed, set some node attributes to avoid side effects with the CSS styles
-                    // which ship with this plugin.
-                } else {
-                    // Change the isexpandable attribute for the parent node to false
-                    // (it's the default in Moodle core, just to be safe).
-                    $customnode->isexpandable = false;
+                // If we are dealing with custom bottom nodes for users and this is the first parent node,
+                // add a special class for later usage in CSS.
+                if ($keyprefix == 'localboostnavigationcustombottomusers' && !$firstcustombuttomusersparentnodeseen) {
+                    $customnode->add_class('localboostnavigationfirstcustombuttomusers');
+                    $firstcustombuttomusersparentnodeseen = true;
+
+                    // Otherwise,
+                    // if we are dealing with custom bottom nodes for admins and this is the first parent node,
+                    // add a special class for later usage in CSS.
+                } else if ($keyprefix == 'localboostnavigationcustombottomadmins' && !$firstcustombuttomadminsparentnodeseen) {
+                    $customnode->add_class('localboostnavigationfirstcustombuttomadmins');
+                    $firstcustombuttomadminsparentnodeseen = true;
                 }
 
                 // Add the custom node to the given navigation_node.
@@ -338,7 +349,7 @@ function local_boostnavigation_build_custom_nodes($customnodes, navigation_node 
                 // The user preference is to collapse the node.
                 if ($userprefcustomnode == 1) {
                     // Set the node to be collapsed.
-                    $customnode->collapse = true;
+                    $customnode->add_class('localboostnavigationcollapsedparent');
 
                     // The user preference is to expand the node.
                 } else {
@@ -348,24 +359,16 @@ function local_boostnavigation_build_custom_nodes($customnodes, navigation_node 
                         // regardless of the user preference.
                         if ($accordionalreadyopen == true) {
                             // Set the node to be collapsed.
-                            $customnode->collapse = true;
+                            $customnode->add_class('localboostnavigationcollapsedparent');
                         } else {
-                            // Set the node to be expanded.
-                            $customnode->collapse = false;
-
                             // If we have set this node to be the expanded node, we must remember this fact for the
                             // remaining accordion nodes.
                             $accordionalreadyopen = true;
                         }
-
-                        // If we don't create an accordion, we can respect the user preference in any case.
-                    } else {
-                        // Set the node to be expanded.
-                        $customnode->collapse = false;
                     }
                 }
 
-                // If the code should be collapsed, remove the active status in any case because otherwise it might get highlighted
+                // If the node should be collapsed, remove the active status in any case because otherwise it might get highlighted
                 // as active which does not make sense for collapse parent nodes.
                 if ($collapse) {
                     $customnode->make_inactive();
@@ -389,8 +392,8 @@ function local_boostnavigation_build_custom_nodes($customnodes, navigation_node 
                     // Remember the node key for collapsing.
                     $collapsenodesforjs[] = $lastparentnode->key;
 
-                    // Change the isexpandable attribute for the parent node to true.
-                    $lastparentnode->isexpandable = true;
+                    // Add the localboostnavigationcollapsibleparent class to the parent node.
+                    $lastparentnode->add_class('localboostnavigationcollapsibleparent');
 
                     // Remember that we have prepared collapsing now.
                     $collapselastparentprepared = true;
@@ -403,8 +406,16 @@ function local_boostnavigation_build_custom_nodes($customnodes, navigation_node 
                 // And change the parent node directly afterwards.
                 $customnode->set_parent($lastparentnode);
 
-                // Set the hidden attribute according to the collapse state of the last parent node.
-                $customnode->hidden = $lastparentnode->collapse;
+                // If the nodes should be collapsed.
+                if ($collapse) {
+                    // Add the localboostnavigationcollapsiblechild class to the child node.
+                    $customnode->add_class('localboostnavigationcollapsiblechild');
+
+                    // And, but only if the last parent node is collapsed, set the child node to be hidden.
+                    if (in_array('localboostnavigationcollapsedparent', $lastparentnode->classes)) {
+                        $customnode->add_class('localboostnavigationcollapsedchild');
+                    }
+                }
 
                 // For some strange reason, Moodle core does only compare the URL base when searching the active navigation node.
                 // This will result in the wrong node being highlighted if we add multiple nodes which only differ by the URL